Move TemplatedSwaggerGenerator to swagger.py
